Update README.md
I just fixed the bulleting for the Team points task.  I changed the _ to * for the unordered list, because my browser was adding the content inline so that the italicized category was on the same line as the preceding text in some cases.  Additioanlly, add a space after the category.